+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com).
 
 ## 0.4.0
 
+### Added
+- Additional ways set y-axis limits.
+
 ### Fixed
 - Updated how fontlist is built to fix [this matplotlib warning](https://github.com/matplotlib/matplotlib/issues/17568).
 - Fixed deprecation warnings.