+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com).
 ## 0.4.0
 
 ### Added
-- Additional ways set y-axis limits. This involves adding `utils.AxLimSetter`, adding the `ylim_setter` param to `draw_logo`
+- Additional ways set y-axis limits. This involves adding `utils.AxLimSetter`, and adding the `ylim_setter` param to `draw_logo` / `draw_line`
+- Allow `show_color` for `draw_line` to be `None`.
 
 ### Fixed
 - Updated how fontlist is built to fix [this matplotlib warning](https://github.com/matplotlib/matplotlib/issues/17568).